% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionGIS.R
\name{AddHoleToPolygon}
\alias{AddHoleToPolygon}
\title{Add holes into Polygons
\code{AddHoleToPolygon}}
\usage{
AddHoleToPolygon(poly, hole)
}
\arguments{
\item{poly}{SpatialPolygons}

\item{hole}{Hole Polygon}
}
\description{
Add holes into Polygons
\code{AddHoleToPolygon}
}
