% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration_CMAES.R
\name{CMAES}
\alias{CMAES}
\title{Automatically calibrate model, that requires user-defined model and objective function.
\code{CMAES}}
\usage{
CMAES(CV, cmd = "pihm++", objfunc = NULL, ...)
}
\arguments{
\item{CV}{Input data, list of data.}

\item{cmd}{Command to run the model.}

\item{objfunc}{User-defined objective function which return the objective values.}

\item{...}{More options passing to objfunc.}
}
\description{
Automatically calibrate model, that requires user-defined model and objective function.
\code{CMAES}
}
