% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionGIS.R
\name{CutSptialLines}
\alias{CutSptialLines}
\title{Cut sptialLines with threshold.
\code{CutSptialLines}}
\usage{
CutSptialLines(sl, tol)
}
\arguments{
\item{sl}{SpatialLines or SpatialLineDataFrame}

\item{tol}{Tolerence. If the length of segment is larger than tolerance, cut the segment until the maximum segment is shorter than tolerance.}
}
\description{
Cut sptialLines with threshold.
\code{CutSptialLines}
}
\examples{
library(PIHMgisR)
library(sp)
x=1:1000/100
l1 = Lines(Line(cbind(x, sin(x)) ), ID='a' )
sl = SpatialLines( list(l1) )
tol1=5;
tol2 =2
sl1 = CutSptialLines(sl, tol1)
sl2 = CutSptialLines(sl, tol2)
par(mfrow=c(1,2))
plot(sl1, col=1:length(sl1));title(paste0('Tol=', tol1))
plot(sl2, col=1:length(sl2));title(paste0('Tol=', tol2))
}
