% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelClasses.R
\docType{data}
\name{.pihm}
\alias{.pihm}
\title{This is .pihm environment.}
\format{An object of class \code{environment} of length 0.}
\usage{
.pihm
}
\description{
This is .pihm environment.
}
\keyword{datasets}
