% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Misc.R
\name{extractCoords}
\alias{extractCoords}
\title{extract Coordinates of SpatialLines or  SpatialPolygons
\code{extractCoords}}
\usage{
extractCoords(x, unique = TRUE, aslist = FALSE)
}
\arguments{
\item{x}{SpatialLines or SpatialPolygons}

\item{unique}{Whether return unique coordinates or duplicated coordinates}

\item{aslist}{Whether return as list of coordinates}
}
\value{
coordinates of points, m x 2.
}
\description{
extract Coordinates of SpatialLines or  SpatialPolygons
\code{extractCoords}
}
\examples{
data('sh')
wb=sh[['wbd']]
p1=extractCoords(wb, aslist = TRUE)
p2=extractCoords(wb, aslist = FALSE)
plot(p2, asp=1)
for(i in 1:length(p1)){
  lines(p1[[i]], col=i+1)
}
}
