% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionGIS.R
\name{extractRaster}
\alias{extractRaster}
\title{\code{extractRaster}}
\usage{
extractRaster(r, xy = NULL, ext = raster::extent(r), plot = T)
}
\arguments{
\item{r}{Raster}

\item{xy}{coordinates of the line, dim=(Npoints, 2); x and y must be in [0, 1]}

\item{ext}{extention of value xy.}
}
\description{
\code{extractRaster}
}
\examples{
library(raster)
}
