% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionGIS.R
\name{fishnet}
\alias{fishnet}
\title{Generatue fishnet
\code{fishnet}}
\usage{
fishnet(ext, crs = sp::CRS("+init=epsg:4326"), dx = diff(ext[1:2])/10,
  dy = dx, lines = FALSE, polygons = TRUE, points = FALSE)
}
\arguments{
\item{ext}{Extension of the fishnet. c(xmin, xmax, ymin, ymax)}

\item{crs}{Projection}

\item{dx}{Interval of x direction}

\item{dy}{Interval of y direction}

\item{lines}{Whether to export SpatialLines}

\item{polygons}{Whether to export SpatialPolygons}

\item{points}{Whether to export SpatialPoints}
}
\description{
Generatue fishnet
\code{fishnet}
}
\examples{
library(raster)
pg=fishnet(ext=c(-80,80, -50,50), dx=5)
plot(pg)
}
