% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Get.R
\name{getCentroid}
\alias{getCentroid}
\title{Get the centroids of the cells
\code{getCentroid}}
\usage{
getCentroid(pm = readmesh())
}
\arguments{
\item{pm}{\code{PIHM.mesh}}
}
\value{
centroids of \code{PIHM.mesh}, ncell x 4 c('x','y','AqD', 'zmax')
}
\description{
Get the centroids of the cells
\code{getCentroid}
}
