% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/River.R
\name{getCriticalRiver}
\alias{getCriticalRiver}
\title{Get Index of critical river channels, that is defined as downstream of river convergence.
\code{getCriticalRiver}}
\usage{
getCriticalRiver(pr = readriv())
}
\arguments{
\item{pr}{PIHM river class}
}
\value{
Index of Critical River Channels, numeric
}
\description{
Get Index of critical river channels, that is defined as downstream of river convergence.
\code{getCriticalRiver}
}
