% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydro_obs.R
\name{LineFit}
\alias{LineFit}
\title{Linefit the observation and simulation
\code{LineFit}}
\usage{
LineFit(x, y = NULL, xlab = "Observation", ylab = "Simulation",
  if.fitline = TRUE, lim, ...)
}
\arguments{
\item{x}{Observation data or matrix/data.frame consist of observation and simulation}

\item{y}{Simulation data}

\item{xlab}{xlab, default is 'Observation'}

\item{ylab}{ylab, default is 'Simulation'}

\item{if.fitline}{Whether plot the fiting line.}

\item{lim}{limits of x and y axis.}

\item{...}{more options for plot()}
}
\value{
Triangle mesh of model domain.
}
\description{
Linefit the observation and simulation
\code{LineFit}
}
\examples{
obs=rnorm(100)
sim=obs+rnorm(100)/2
LineFit(cbind(obs,sim))
}
