% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.DomainDecomposition.R
\name{m.DomainDecomposition}
\alias{m.DomainDecomposition}
\title{Generate triangle mesh domain
\code{m.DomainDecomposition}}
\usage{
m.DomainDecomposition(wb, dem = NULL, riv = NULL, lk = NULL,
  pts = NULL, q = 30, ...)
}
\arguments{
\item{wb}{SpatialPolygon or SpatialLines which define the watershed boundary}

\item{dem}{Elevation data.}

\item{riv}{SpatialLines of river network, optional}

\item{lk}{SpatialPolygon of lake.}

\item{q}{minimum angle of triangle}

\item{...}{more options in RTriangle::triangulate()}
}
\value{
Coordinates of triangles centroids
}
\description{
Generate triangle mesh domain
\code{m.DomainDecomposition}
}
