% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionGIS.R
\name{MeshData2Raster}
\alias{MeshData2Raster}
\title{SpatialData to Raster
\code{MeshData2Raster}}
\usage{
MeshData2Raster(x = getElevation(), rmask = PIHM.mask(proj = proj),
  pm = readmesh(), proj = NULL, stack = FALSE, plot = FALSE)
}
\arguments{
\item{x}{vector or matrix, length/nrow is number of cells.}

\item{rmask}{mask of PIHM mesh}

\item{pm}{PIHM mesh}

\item{proj}{Projejction parameter}

\item{stack}{Whether export the stack, only when the x is a matrix, i.e. (Ntime x Ncell).}
}
\value{
Raster map
}
\description{
SpatialData to Raster
\code{MeshData2Raster}
}
