% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInfo.R
\name{ModelInfo}
\alias{ModelInfo}
\title{Summary of the basic information of PIHM input data.
\code{ModelInfo}}
\usage{
ModelInfo(path = get("outpath", envir = .pihm))
}
\arguments{
\item{path}{Where to export the basic model infomation.}
}
\value{
Basic model infomation, figures and tables
}
\description{
Summary of the basic information of PIHM input data.
\code{ModelInfo}
}
