% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelClasses.R
\docType{class}
\name{PIHM mesh-class}
\alias{PIHM mesh-class}
\alias{PIHM.mesh}
\title{Class of PIHM.mesh
\code{PIHM.mesh}}
\value{
Class of PIHM.mesh
}
\description{
Class of PIHM.mesh
\code{PIHM.mesh}
}
\section{Slots}{

\describe{
\item{\code{mesh}}{data.frame or matrix}

\item{\code{point}}{data.frame or matrix}
}}

