% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelClasses.R
\docType{class}
\name{PIHM River-class}
\alias{PIHM River-class}
\alias{PIHM.river}
\title{Class of PIHM.river
\code{PIHM.river}}
\value{
Class of PIHM.river
}
\description{
Class of PIHM.river
\code{PIHM.river}
}
\section{Slots}{

\describe{
\item{\code{river}}{data.frame or matrix}

\item{\code{rivertype}}{data.frame or matrix}

\item{\code{point}}{data.frame or matrix}
}}

