% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{PIHM.filein}
\alias{PIHM.filein}
\title{Prepare Input file of PIHM
\code{PIHM.filein}}
\usage{
PIHM.filein(projname = get("PRJNAME", envir = .pihm),
  inpath = get("inpath", envir = .pihm), outpath = get("outpath", envir
  = .pihm))
}
\arguments{
\item{projname}{Character, project name, default= PRJNAME which is a global variable.}

\item{inpath}{PIHM input directory, default = inpath which is global variable}

\item{outpath}{PIHM output directory, default = outpath which is global variable}
}
\value{
Character of full path of input files for PIHM
}
\description{
Prepare Input file of PIHM
\code{PIHM.filein}
}
