% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelConfigure.R
\name{pihm.init}
\alias{pihm.init}
\title{extract Coordinates of SpatialLines or  SpatialPolygons
\code{pihm.init}}
\usage{
pihm.init(ncell, nriv, AqD = 10, stage = 0.1, p1 = 0.4, p2 = p1)
}
\arguments{
\item{ncell}{Number of triangles in PIHM mesh domain}

\item{nriv}{Number of river lines}

\item{AqD}{Default aquifer depth}

\item{stage}{Default river stage}

\item{p1}{Fraction of unsat zone, [0,1]}

\item{p2}{Fraction of Groundwater, [0,1]}
}
\value{
List of initial condition
}
\description{
extract Coordinates of SpatialLines or  SpatialPolygons
\code{pihm.init}
}
