% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{PIHM.mask}
\alias{PIHM.mask}
\title{Generate the raster mask of Mesh domain
\code{PIHM.mask}}
\usage{
PIHM.mask(pm = readmesh(), proj = NULL, rr = get("PIHM.MASK", envir =
  .pihm), ngrids = 200, resolution = NULL)
}
\arguments{
\item{pm}{\code{PIHM.mesh}}

\item{proj}{Projection parameter}

\item{rr}{Default mask in .pihm environment}

\item{ngrids}{Number of grid along x direction.}

\item{resolution}{Resolution, defaul = NULL, resolution = extent / ngrids}
}
\value{
Raster map
}
\description{
Generate the raster mask of Mesh domain
\code{PIHM.mask}
}
