% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIHMgisEnv.R
\name{PIHM}
\alias{PIHM}
\title{Setup .pihm environment.
\code{read.tsd}}
\usage{
PIHM(prjname, inpath = file.path("input", prjname),
  outpath = file.path("output", paste0(prjname, ".out")),
  anapath = file.path(outpath, "PIHManalysis"))
}
\arguments{
\item{prjname}{charactor of PIHM project name.}

\item{inpath}{path of PIHM input directory}

\item{outpath}{path of PIHM output directory}

\item{anapath}{charactor of PIHM project name.}
}
\description{
Setup .pihm environment.
\code{read.tsd}
}
