% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeshDomain.R
\name{pihmAtt}
\alias{pihmAtt}
\title{extract Coordinates of SpatialLines or  SpatialPolygons
\code{pihmAtt}}
\usage{
pihmAtt(tri, r.soil = NULL, r.geol = NULL, r.lc = NULL,
  r.forc = NULL, r.mf = NULL, r.bc = NULL)
}
\arguments{
\item{tri}{Triangles defination}

\item{r.soil}{raster of soil classification}

\item{r.geol}{raster of geology layer}

\item{r.lc}{raster of land cover, LAI and Roughness Length}

\item{r.forc}{raster of forcing data}

\item{r.mf}{raster of melt factor}

\item{r.bc}{raster of boundary condition}
}
\value{
data.frame of PIHM .att
}
\description{
extract Coordinates of SpatialLines or  SpatialPolygons
\code{pihmAtt}
}
