% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelConfigure.R
\name{pihmcalib}
\alias{pihmcalib}
\title{Generate the default PIHM model calibration
\code{pihmcalib}}
\usage{
pihmcalib()
}
\value{
Default calibration values for PIHM
}
\description{
Generate the default PIHM model calibration
\code{pihmcalib}
}
