% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelConfigure.R
\name{pihmpara}
\alias{pihmpara}
\title{Generate the default PIHM model configuration
\code{pihm.para}}
\usage{
pihmpara(nday = 10)
}
\arguments{
\item{nday}{Simulation periods (days)}
}
\value{
model configureation, Vector
}
\description{
Generate the default PIHM model configuration
\code{pihm.para}
}
