% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/River.R
\name{pihmRiver}
\alias{pihmRiver}
\title{calculate river order, downstream, slope and length
\code{pihmRiver}}
\usage{
pihmRiver(sl, dem)
}
\arguments{
\item{sl}{SpatialLines*}

\item{dem}{Raster of elevation}
}
\value{
PIHM river
}
\description{
calculate river order, downstream, slope and length
\code{pihmRiver}
}
