% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rvier2.R
\name{pihmRiver2}
\alias{pihmRiver2}
\title{calculate river order
\code{pihmRiver2}}
\usage{
pihmRiver2(sp, dem, minslope = 1/1000)
}
\arguments{
\item{sp}{SpatialLines*}

\item{dem}{Raster of elevation}

\item{minslope}{Miminum slope of river bed.}
}
\value{
Stream Order of SpatialLines*
}
\description{
calculate river order
\code{pihmRiver2}
}
