% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/River.R
\name{pihmRiverSeg}
\alias{pihmRiverSeg}
\title{Generate the river segments table
\code{pihmRiverSeg}}
\usage{
pihmRiverSeg(sl)
}
\arguments{
\item{sl}{SpatialLinesDataFrame}
}
\value{
river segments table
}
\description{
Generate the river segments table
\code{pihmRiverSeg}
}
