% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionGIS.R
\name{PointInDistance}
\alias{PointInDistance}
\title{Simplify SpatialData.
\code{PointInDistance}}
\usage{
PointInDistance(pt, tol)
}
\arguments{
\item{tol}{Tolerance}

\item{x}{2-column}
}
\value{
Index of points that within tol
}
\description{
Simplify SpatialData.
\code{PointInDistance}
}
