% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinput.R
\name{readpc}
\alias{readpc}
\title{Read the .para or .calib file
\code{readpc}}
\usage{
readpc(file = PIHM.filein()["md.para"])
}
\arguments{
\item{file}{full path of file}
}
\value{
.para or .calib
}
\description{
Read the .para or .calib file
\code{readpc}
}
