% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinput.R
\name{readriv}
\alias{readriv}
\title{Read the .riv file
\code{readriv}}
\usage{
readriv(file = PIHM.filein()["md.riv"])
}
\arguments{
\item{file}{full path of .riv file}
}
\value{
pr \code{PIHM.riv}
}
\description{
Read the .riv file
\code{readriv}
}
