% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rect.ts}
\alias{rect.ts}
\title{Find the rect for banded background.}
\usage{
rect.ts(x, time.col = "year")
}
\arguments{
\item{x}{xts data}

\item{time.col}{Time interval of banded background. Default is 'year'}
}
\description{
Find the rect for banded background.
}
