% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rowMatch}
\alias{rowMatch}
\title{Identify whether every row of \code{m} is same as \code{v};}
\usage{
rowMatch(x, m)
}
\arguments{
\item{x}{required}

\item{m}{required}
}
\value{
TRUE/FALSE of matching
}
\description{
Identify whether every row of \code{m} is same as \code{v};
}
