% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionGIS.R
\name{SimpleSpatial}
\alias{SimpleSpatial}
\title{Simplify SpatialData.
\code{SimpleSpatial}}
\usage{
SimpleSpatial(x)
}
\arguments{
\item{x}{SpatialData}
}
\value{
Simplified SpatialData
}
\description{
Simplify SpatialData.
\code{SimpleSpatial}
}
