% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_RiverProcess.R
\name{sp.DissolveLines}
\alias{sp.DissolveLines}
\title{Dissolve River segments
\code{sp.RiverDissolve}}
\usage{
sp.DissolveLines(sp)
}
\arguments{
\item{sp}{SpatialLines}
}
\value{
SpatialLinesDataFrame
}
\description{
Dissolve River segments
\code{sp.RiverDissolve}
}
