% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_RiverProcess.R
\name{sp.RiverPath}
\alias{sp.RiverPath}
\title{determine River path
\code{sp.RiverPath}}
\usage{
sp.RiverPath(sp, idown = sp.RiverDown(sp, coord = coord))
}
\arguments{
\item{sp}{SpatialLines*}

\item{idown}{coordinates of \code{sp}.}
}
\value{
List of River Path
}
\description{
determine River path
\code{sp.RiverPath}
}
