% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterBalance.R
\name{wb.ele}
\alias{wb.ele}
\title{Calculate the water balance
\code{wb.riv}}
\usage{
wb.ele(xl = BasicPlot(varname = c(paste0("elev", c("prcp", "et0", "et1",
  "et2"), "eleq", c("surf", "sub"), "eley", c("surf", "unsat", "gw")))),
  fun = xts::apply.yearly, period = "years", plot = TRUE)
}
\arguments{
\item{xl}{List of data. Five variables are included: prcp (eleqprcp), et0 (eleqet0), et1 (eleqet1), et2 (eleqet2) and discharge (rivqflx)}

\item{fun}{function to process the time-series data. Default = apply.daily.}

\item{period}{Period of the waterbalance}

\item{plot}{Whether plot the result}
}
\value{
A matrix, contains the colums of water balance factors
}
\description{
Calculate the water balance
\code{wb.riv}
}
