% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{which.row}
\alias{which.row}
\title{Find the match row in a matrix
\code{which.row} Which row in Matrix \code{mat} is identical to \code{v}}
\usage{
which.row(v, mat)
}
\arguments{
\item{v}{a vector}

\item{mat}{a matrix}
}
\description{
Find the match row in a matrix
\code{which.row} Which row in Matrix \code{mat} is identical to \code{v}
}
\examples{
x = c(1,2)
mat = t(matrix(c(1,2, 2,3, 1,2, 4,5), ncol=2))
which.row(x, mat)
}
