% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePIHM.R
\name{write.config}
\alias{write.config}
\title{Write PIHM configuration files (.para, .calib, etc.)}
\usage{
write.config(x, file, backup = TRUE)
}
\arguments{
\item{x}{PIHM model configure parameter or calibration}

\item{file}{file name}

\item{backup}{TRUE/FALSE}
}
\description{
Write PIHM configuration files (.para, .calib, etc.)
}
