% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePIHM.R
\name{write.pc}
\alias{write.pc}
\title{Write PIHM .para or .calib file
\code{write.pc}}
\usage{
write.pc(x, file, backup = TRUE)
}
\arguments{
\item{x}{PIHM model configure parameter or calibration}

\item{file}{file name}

\item{backup}{TRUE/FALSE}
}
\description{
Write PIHM .para or .calib file
\code{write.pc}
}
