% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePIHM.R
\name{writeforc}
\alias{writeforc}
\title{Write PIHM .forc file
\code{writeforc}}
\usage{
writeforc(fns, path = "", startdate = "20000101", file)
}
\arguments{
\item{fns}{filenames; Vector of character,}

\item{path}{Common path of the files.}

\item{startdate}{Start Date. Character. e.g. 20000101}

\item{file}{file name}
}
\description{
Write PIHM .forc file
\code{writeforc}
}
