% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePIHM.R
\name{writemesh}
\alias{writemesh}
\title{Write PIHM .mesh file
\code{writemesh}}
\usage{
writemesh(pm, file, backup = TRUE)
}
\arguments{
\item{pm}{PIHMmesh class}

\item{file}{file name}

\item{backup}{TRUE/FALSE, if backup the existing file}
}
\description{
Write PIHM .mesh file
\code{writemesh}
}
