% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionGIS.R
\name{extractRaster}
\alias{extractRaster}
\title{\code{extractRaster}}
\usage{
extractRaster(r, s = 0, loc = c(0.5, 0.5), ext = raster::extent(r),
  dx = raster::res(r)[1])
}
\arguments{
\item{r}{Raster}

\item{s}{Slope of the line}

\item{loc}{ratio on axis for (x,y), (0.5,0.5) is the center of plot}

\item{ext}{extention of plot}

\item{dx}{resolution}
}
\description{
\code{extractRaster}
}
\examples{
library(raster)
}
