% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GreemAmpt.R
\name{GreenAmpt}
\alias{GreenAmpt}
\title{Calculate infiltration with Green Ampt method
\code{GreenAmpt}}
\usage{
GreenAmpt(theta, thetas, ksat, Hfront, y0 = 0.01, Time = 0)
}
\arguments{
\item{theta}{Soil moisture content}

\item{thetas}{Saturated moisture content}

\item{ksat}{Saturated hydraulic conductivity}

\item{Hfront}{Depth of wetting front.}

\item{y0}{Depth of ponding water.}

\item{Time}{Time.}
}
\value{
Infiltration rate
}
\description{
Calculate infiltration with Green Ampt method
\code{GreenAmpt}
}
\references{
Green, W.H. and G. Ampt. 1911. Studies of soil physics, part I – the flow of air and water through soils. J. Ag. Sci. 4:1-24.
}
