% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{PIHM.mask}
\alias{PIHM.mask}
\title{Generate the raster mask of Mesh domain
\code{PIHM.mask}}
\usage{
PIHM.mask(pm = readmesh(), crs = NULL, rr = get("PIHM.MASK", envir =
  .pihm), n = 40000, cellsize = NULL)
}
\arguments{
\item{pm}{\code{PIHM.mesh}}

\item{crs}{Projection parameter}

\item{rr}{Default mask in .pihm environment}

\item{n}{Number of grid}

\item{cellsize}{Resolution, defaul = NULL, resolution = extent / ngrids}
}
\value{
Raster map
}
\description{
Generate the raster mask of Mesh domain
\code{PIHM.mask}
}
