% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePIHM.R
\name{writeinit}
\alias{writeinit}
\title{Write PIHM .ic file
\code{writeinit}}
\usage{
writeinit(x, file)
}
\arguments{
\item{x}{Initial condition, list()}

\item{file}{file name}
}
\description{
Write PIHM .ic file
\code{writeinit}
}
