% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePIHM.R
\name{writeriv}
\alias{writeriv}
\title{Write PIHM .mesh file
\code{writemesh}}
\usage{
writeriv(pr, file, backup = TRUE)
}
\arguments{
\item{pr}{PIHMmesh class}

\item{file}{file name}

\item{backup}{TRUE/FALSE, if backup the existing file}
}
\description{
Write PIHM .mesh file
\code{writemesh}
}
